# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ethree API",
    "version": "1.0.0",
    "description": "API for Ethree - Eat, Enjoy, Entertainment platform (Supabase PostgreSQL Version)"
  },
  "servers": [
    {
      "url": "https://e3-e4-backend.vercel.app",
      "description": "Production Server"
    },
    {
      "url": "http://localhost:5000",
      "description": "Local Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/e3/rides": {
      "get": {
        "summary": "Get all E3 rides",
        "tags": [
          "E3"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "all",
            "schema": {
              "type": "string"
            },
            "description": "Pass true to fetch all rides including offline/inactive rides"
          }
        ],
        "responses": {
          "200": {
            "description": "List of E3 rides",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ride"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new ride (Admin only)",
        "tags": [
          "E3"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Ride"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ride created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ride"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "403": {
            "description": "Admin access required"
          }
        }
      }
    },
    "/api/e3/dine": {
      "get": {
        "summary": "Get all E3 dine items",
        "tags": [
          "E3"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "all",
            "schema": {
              "type": "string"
            },
            "description": "Pass true to fetch all dine items including closed/inactive ones"
          }
        ],
        "responses": {
          "200": {
            "description": "List of E3 dine items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DineItem"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new dine item (Admin only)",
        "tags": [
          "E3"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DineItem"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dine item created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DineItem"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "403": {
            "description": "Admin access required"
          }
        }
      }
    },
    "/api/e3/rides/{id}": {
      "put": {
        "summary": "Update a ride (Admin only)",
        "tags": [
          "E3"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Ride ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Ride"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ride updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ride"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "403": {
            "description": "Admin access required"
          },
          "404": {
            "description": "Ride not found"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a ride (Admin only)",
        "tags": [
          "E3"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ride deleted"
          },
          "403": {
            "description": "Admin access required"
          },
          "404": {
            "description": "Ride not found"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/e3/dine/{id}": {
      "put": {
        "summary": "Update a dine item (Admin only)",
        "tags": [
          "E3"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Dine item ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DineItem"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dine item updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DineItem"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "403": {
            "description": "Admin access required"
          },
          "404": {
            "description": "Dine item not found"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a dine item (Admin only)",
        "tags": [
          "E3"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dine item deleted"
          },
          "403": {
            "description": "Admin access required"
          },
          "404": {
            "description": "Dine item not found"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}